+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/Chicken_Attack.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: Chicken_Attack
   m_EditorClassIdentifier: 
   wordToUse: "\u30DD\u30A4\u30C3"
+  onomatoSprite: {fileID: 21300000, guid: 971442f7dd37c504da62d32a77293eea, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: f0c6b01d5911abd4e8ba109c8ef657b9, type: 2}
   fontAsset: {fileID: 11400000, guid: 40ebecea6a7fb6b4e9bfeb174c259461, type: 2}
   fontColor:
     serializedVersion: 2
