+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/Hunter_Walk.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: Hunter_Walk
   m_EditorClassIdentifier: 
   wordToUse: "\u30BA\u30B7\u30BA\u30B7"
+  onomatoSprite: {fileID: 21300000, guid: 18da527d85ae57140aa486dac770465a, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: 7d23c146b79a6da4399a34659674a5be, type: 2}
   fontAsset: {fileID: 11400000, guid: ad584379139b1df46b9777099a13e04e, type: 2}
   fontColor:
     serializedVersion: 2
