+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/Knight_Attack.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: Knight_Attack
   m_EditorClassIdentifier: 
   wordToUse: "\u30B7\u30E5\u30C3"
+  onomatoSprite: {fileID: 21300000, guid: 1940f868511447e4ca131694313f42f4, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: 668ae3d433d9e9242b2340e00a8472ce, type: 2}
   fontAsset: {fileID: 11400000, guid: ad584379139b1df46b9777099a13e04e, type: 2}
   fontColor:
     serializedVersion: 2
