+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/Knight_Walk.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: Knight_Walk
   m_EditorClassIdentifier: 
   wordToUse: "\u30AC\u30B7\u30E3"
+  onomatoSprite: {fileID: 21300000, guid: 24d456c2b5e364e428c99660154bf6e4, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: e76353bfdbbe4b64dadfb020ea55019b, type: 2}
   fontAsset: {fileID: 11400000, guid: 178211434d9d2d14fa419a5f9dd45ba6, type: 2}
   fontColor:
     serializedVersion: 2
