+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/UFO_Attack.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: UFO_Attack
   m_EditorClassIdentifier: 
   wordToUse: "\u30D6\u30A5\u30A6\u30F3"
+  onomatoSprite: {fileID: 21300000, guid: e4e47f1fd17cac44393530fa8a40d307, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: bf1d2a129097b634b842b412432e0f44, type: 2}
   fontAsset: {fileID: 11400000, guid: ad584379139b1df46b9777099a13e04e, type: 2}
   fontColor:
     serializedVersion: 2
