+++ MS_Project/Assets/Data/Onomatopoeia/Enemy_Onomatope/Virus_Attack.asset
@@ -13,6 +13,9 @@ MonoBehaviour:
   m_Name: Virus_Attack
   m_EditorClassIdentifier: 
   wordToUse: "\u30D6\u30E8\u30C3"
+  onomatoSprite: {fileID: 21300000, guid: 522f027afda883b4ea218bf4df675c0a, type: 3}
+  spriteSize: 1
+  onomatoACont: {fileID: 22100000, guid: 71c87be03ca95bb4d959d3eace749d73, type: 2}
   fontAsset: {fileID: 11400000, guid: ad584379139b1df46b9777099a13e04e, type: 2}
   fontColor:
     serializedVersion: 2
