--- MS_Project/Assets/Scripts/UI/Pause/GameCanvas.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: db66321d7f8c2b54b9ed712a2477ed39
