--- MS_Project/Assets/Scripts/UI/Pause/PauseGame.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c91a6621379f1fe4f9bc9152239f5f93
