--- MS_Project/Assets/Scripts/UI/Pause/RestartGame.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class RestartGame : MonoBehaviour
-{
-    void Update()
-    {
-            Time.timeScale = 1;
-    }
-}
