--- MS_Project/Assets/Scripts/UI/Pause/RestartGame.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a779120da6a8d6f4eb73bdf44f37be77
