+++ MS_Project/Assets/Scripts/UI/Pause/SamplePause.unity
@@ -133,7 +133,6 @@ GameObject:
   - component: {fileID: 62486189}
   - component: {fileID: 62486193}
   - component: {fileID: 62486192}
-  - component: {fileID: 62486194}
   m_Layer: 0
   m_Name: Game
   m_TagString: Untagged
@@ -246,18 +245,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: c3beb83fc10fc9d47a519f1ec56ac9b2, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &62486194
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 62486186}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a779120da6a8d6f4eb73bdf44f37be77, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &464389910
@@ -740,7 +727,6 @@ GameObject:
   - component: {fileID: 1807126593}
   - component: {fileID: 1807126592}
   - component: {fileID: 1807126595}
-  - component: {fileID: 1807126597}
   m_Layer: 0
   m_Name: Pause
   m_TagString: Untagged
@@ -844,18 +830,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   canvas: {fileID: 62486186}
---- !u!114 &1807126597
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1807126590}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c91a6621379f1fe4f9bc9152239f5f93, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1807206759
 GameObject:
   m_ObjectHideFlags: 0
